backend/requirements.txt
Flask==3.0.3
flask-marshmallow==1.2.1
Flask-SQLAlchemy==3.1.1
marshmallow-sqlalchemy==1.1.0
